+++ package.yaml
@@ -115,12 +115,14 @@ library:
     - Juvix.Encoding.Mendler
     - Juvix.Encoding.Scott
     - Juvix.Encoding.Types
-    - Juvix.LLVM.Codegen.Constants
-    - Juvix.LLVM.Codegen.Sum
-    - Juvix.LLVM.Codegen.Types
-    - Juvix.LLVM.Codegen.Shared
-    - Juvix.LLVM.Codegen.Block
-    - Juvix.LLVM.Codegen.Graph
+    - Juvix.Backends
+    - Juvix.Backends.LLVM
+    - Juvix.Backends.LLVM.Codegen.Constants
+    - Juvix.Backends.LLVM.Codegen.Sum
+    - Juvix.Backends.LLVM.Codegen.Types
+    - Juvix.Backends.LLVM.Codegen.Shared
+    - Juvix.Backends.LLVM.Codegen.Block
+    - Juvix.Backends.LLVM.Codegen.Graph
   other-modules:
     - Juvix.Interpreter.InteractionNet.NodeInterface
     - Juvix.Utility.Helper
