+++ src/Juvix/Backends/LLVM/Codegen/Graph.hs
@@ -1,8 +1,8 @@
 -- | Operations necessary to update nodes
-module Juvix.LLVM.Codegen.Graph where
+module Juvix.Backends.LLVM.Codegen.Graph where
 
-import Juvix.LLVM.Codegen.Block as Block
-import Juvix.LLVM.Codegen.Types as Types
+import Juvix.Backends.LLVM.Codegen.Block as Block
+import Juvix.Backends.LLVM.Codegen.Types as Types
 import Juvix.Library hiding (Type, local)
 import qualified Juvix.Utility.HashMap as Map
 import qualified LLVM.AST as AST
