+++ src/Juvix/Backends/LLVM/Codegen/Shared.hs
@@ -1,5 +1,5 @@
 -- | Shared between Types and Sum
-module Juvix.LLVM.Codegen.Shared where
+module Juvix.Backends.LLVM.Codegen.Shared where
 
 import Juvix.Library hiding (Type)
 import qualified Juvix.Utility.HashMap as Map
